Migrate to null-safety, upgrade dependencies, add Web support.

+++ lib/models/named_data_store.dart
@@ -9,40 +9,43 @@ class NamedDataStore {
 
   LinkedHashMap<String, String> _entries = LinkedHashMap();
 
-  File _registryFile;
-
   final Uuid uuid = Uuid();
 
   NamedDataStore(this.storeName);
 
-  Future<Map> readRegistry() async {
+  Future<Map<String, String>> readRegistry() async {
     _entries.clear();
 
-    Directory directory = await getApplicationDocumentsDirectory();
+    Directory? directory = await getApplicationSupportDirectory();
 
-    _registryFile = File('${directory.path}/$storeName/registry.txt');
+    if (directory != null) {
+      File _registryFile = File('${directory.path}/$storeName/registry.txt');
 
-    if (!await _registryFile.exists()) {
-      await _registryFile.create(recursive: true);
+      if (!await _registryFile.exists()) {
+        await _registryFile.create(recursive: true);
+        return _entries;
+      }
+
+      String contents = await _registryFile.readAsString();
+
+      for (String entry in contents.split('\n')) {
+        int separator = entry.indexOf(';');
+
+        if (separator == -1) {
+          if (entry.isNotEmpty) {
+            print('Ignoring invalid entry: $entry');
+          }
+          continue;
+        }
+
+        _entries[entry.substring(0, separator)] =
+            entry.substring(separator + 1);
+      }
+
       return _entries;
     }
 
-    String contents = await _registryFile.readAsString();
-
-    for (String entry in contents.split('\n')) {
-      int separator = entry.indexOf(';');
-
-      if (separator == -1) {
-        if (entry.isNotEmpty) {
-          print('Ignoring invalid entry: $entry');
-        }
-        continue;
-      }
-
-      _entries[entry.substring(0, separator)] = entry.substring(separator + 1);
-    }
-
-    return _entries;
+    return Map();
   }
 
   Future<bool> _writeRegistry() async {
@@ -51,39 +54,45 @@ class NamedDataStore {
     _entries
         .forEach((String uuid, String name) => stringValue += '$uuid;$name\n');
 
-    Directory directory = await getApplicationDocumentsDirectory();
+    Directory? directory = await getApplicationDocumentsDirectory();
 
-    _registryFile = File('${directory.path}/$storeName/registry.txt');
-
-    if (!await _registryFile.exists()) {
-      await _registryFile.create(recursive: true);
+    if (directory != null) {
+      File _registryFile = File('${directory.path}/$storeName/registry.txt');
 
       if (!await _registryFile.exists()) {
-        return false;
+        await _registryFile.create(recursive: true);
+
+        if (!await _registryFile.exists()) {
+          return false;
+        }
       }
+
+      await _registryFile.writeAsString(stringValue);
+      return true;
     }
 
-    await _registryFile.writeAsString(stringValue);
-
-    return true;
+    return false;
   }
 
   Future<bool> _writeData(String uuid, String data) async {
-    Directory directory = await getApplicationDocumentsDirectory();
+    Directory? directory = await getApplicationDocumentsDirectory();
 
-    File dataFile = File('${directory.path}/$storeName/$uuid.txt');
-
-    if (!await dataFile.exists()) {
-      await dataFile.create();
+    if (directory != null) {
+      File dataFile = File('${directory.path}/$storeName/$uuid.txt');
 
       if (!await dataFile.exists()) {
-        return false;
+        await dataFile.create();
+
+        if (!await dataFile.exists()) {
+          return false;
+        }
       }
+
+      await dataFile.writeAsString(data);
+      return true;
     }
 
-    await dataFile.writeAsString(data);
-
-    return true;
+    return false;
   }
 
   Future<String> saveNewData(String name, String data) async {
@@ -106,9 +115,6 @@ class NamedDataStore {
     await readRegistry();
 
     if (_entries.containsKey(uuid) && await _writeData(uuid, data)) {
-      String name = _entries[uuid];
-      _entries.remove(uuid);
-      _entries[uuid] = name;
       await _writeRegistry();
       return true;
     }
@@ -130,27 +136,31 @@ class NamedDataStore {
 
   Future<String> readData(String uuid) async {
     if (_entries.containsKey(uuid)) {
-      Directory directory = await getApplicationDocumentsDirectory();
+      Directory? directory = await getApplicationDocumentsDirectory();
 
-      File dataFile = File('${directory.path}/$storeName/$uuid.txt');
+      if (directory != null) {
+        File dataFile = File('${directory.path}/$storeName/$uuid.txt');
 
-      if (await dataFile.exists()) {
-        return dataFile.readAsString();
+        if (await dataFile.exists()) {
+          return dataFile.readAsString();
+        }
       }
     }
 
-    return null; // TODO Better error handling
+    return ''; // TODO Better error handling
   }
 
   Future<bool> deleteData(String uuid) async {
     if (_entries.containsKey(uuid)) {
-      Directory directory = await getApplicationDocumentsDirectory();
+      Directory? directory = await getApplicationDocumentsDirectory();
 
-      await File('${directory.path}/$storeName/$uuid.txt').delete();
+      if (directory != null) {
+        await File('${directory.path}/$storeName/$uuid.txt').delete();
 
-      _entries.remove(uuid);
-      await _writeRegistry();
-      return true;
+        _entries.remove(uuid);
+        await _writeRegistry();
+        return true;
+      }
     }
 
     return false;
